requirements.txt
backports-abc==0.4
beautifulsoup4==4.5.1
bs4==0.0.1
certifi==2016.8.8
elasticsearch==2.4.0
greenlet==0.4.10
motor==0.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pymongo==2.8
pynsq==0.7.0
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sleekxmpp==1.3.1
tornado==4.4.1
torsession==0.2.9
urllib3==1.16
WTForms==2.1
wtforms-tornado==0.0.2
